added initial comm config
