+++ src/main/kotlin/madalv/models/Config.kt
@@ -1,4 +1,4 @@
-package madalv
+package madalv.models
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
