--- src/main/kotlin/madalv/plugins/Sockets.kt
@@ -1,110 +1,0 @@
-package madalv.plugins
-
-import io.ktor.network.selector.*
-import io.ktor.network.sockets.*
-import io.ktor.utils.io.*
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.runBlocking
-import java.io.InputStream
-import java.util.*
-import io.ktor.server.routing.*
-import io.ktor.server.websocket.*
-import io.ktor.websocket.*
-import java.time.Duration
-import io.ktor.server.application.*
-import io.ktor.server.response.*
-import io.ktor.server.request.*
-import io.ktor.server.routing.*
-import madalv.cfg
-
-fun Application.configureSockets() {
-
-    install(WebSockets) {
-        pingPeriod = Duration.ofSeconds(15)
-        timeout = Duration.ofSeconds(15)
-        maxFrameSize = Long.MAX_VALUE
-        masking = false
-    }
-
-    routing {
-        webSocket("/ws") { // websocketSession
-            for (frame in incoming) {
-                if (frame is Frame.Text) {
-                    val text = frame.readText()
-                    outgoing.send(Frame.Text("YOU SAID: $text"))
-                    if (text.equals("bye", ignoreCase = true)) {
-                        close(CloseReason(CloseReason.Codes.NORMAL, "Client said BYE"))
-                    }
-                }
-            }
-        }
-    }
-}
-
-/**
- * Two mains are provided, you must first start EchoApp.Server, and then EchoApp.Client.
- * You can also start EchoApp.Server and then use a telnet client to connect to the echo server.
- */
-object EchoApp {
-    val selectorManager = ActorSelectorManager(Dispatchers.IO)
-    val DefaultPort = cfg.tcpPort
-
-    object Server {
-        @JvmStatic
-        fun main(args: Array<String>) {
-            runBlocking {
-                val serverSocket = aSocket(selectorManager).tcp().bind(port = DefaultPort)
-                println("Echo Server listening at ${serverSocket.localAddress}")
-                while (true) {
-                    val socket = serverSocket.accept()
-                    println("Accepted $socket")
-                    launch {
-                        val read = socket.openReadChannel()
-                        val write = socket.openWriteChannel(autoFlush = true)
-                        try {
-                            while (true) {
-                                val line = read.readUTF8Line()
-                                write.writeStringUtf8("$line\n")
-                            }
-                        } catch (e: Throwable) {
-                            socket.close()
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    object Client {
-        @JvmStatic
-        fun main(args: Array<String>) {
-            runBlocking {
-                val socket = aSocket(selectorManager).tcp().connect("127.0.0.1", port = DefaultPort)
-                val read = socket.openReadChannel()
-                val write = socket.openWriteChannel(autoFlush = true)
-
-                launch(Dispatchers.IO) {
-                    while (true) {
-                        val line = read.readUTF8Line()
-                        println("server: $line")
-                    }
-                }
-
-                for (line in System.`in`.lines()) {
-                    println("client: $line")
-                    write.writeStringUtf8("$line\n")
-                }
-            }
-        }
-
-        private fun InputStream.lines() = Scanner(this).lines()
-
-        private fun Scanner.lines() = sequence {
-            while (hasNext()) {
-                yield(readLine())
-            }
-        }
-    }
-}
-
